Extract geolocation to main component
For starters, let's drop the Navigator component since it was a bad
design choice :smile:

Also, let's rename Forecast to ForecastView in order to improve
organization.

Finally, this commit also adds a basic LoadingView that can be displayed
when fetching data from the web.
